test new behavior with maven plugin

+++ org.kevoree.modeling.generator.mavenplugin/src/main/java/org/kevoree/modeling/generator/mavenplugin/HtmlTemplateGenerator.java
@@ -12,35 +12,32 @@ public class HtmlTemplateGenerator {
 
 
     public static void generateHtml(Path targetDir, String jsFileName, String metaModelFQN) throws IOException {
-        String  metaModelName;
-        if(metaModelFQN.contains(".")) {
-           metaModelName = metaModelFQN.substring(metaModelFQN.lastIndexOf(".")+1);
+        String metaModelName;
+        if (metaModelFQN.contains(".")) {
+            metaModelName = metaModelFQN.substring(metaModelFQN.lastIndexOf(".") + 1);
         } else {
             metaModelName = metaModelFQN;
             metaModelFQN = metaModelFQN.toLowerCase() + "." + metaModelFQN;
         }
 
-
         StringBuilder sb = new StringBuilder();
         sb.append("<!DOCTYPE html>\n<html>\n<head lang=\"en\">\n    <meta charset=\"UTF-8\">\n    <title></title>\n</head>\n");
         sb.append("<body>\n");
-        sb.append("<script src=\"./java.js\"></script>\n");
-        sb.append("<script src=\"./org.kevoree.modeling.microframework.typescript.js\"></script>\n");
-        sb.append("<script src=\"./"+jsFileName+"\"></script>\n");
+        sb.append("<script src=\"./" + jsFileName + "\"></script>\n");
         sb.append("<script>\n");
-        sb.append("    var "+metaModelName.toLowerCase()+"Universe = new "+metaModelFQN+"Universe();\n");
-        sb.append("    "+metaModelName.toLowerCase()+"Universe.newDimension(function(dimension){\n");
-        sb.append("        var "+metaModelName.toLowerCase()+"View = dimension.time(0);\n");
+        sb.append("    var " + metaModelName.toLowerCase() + "Universe = new " + metaModelFQN + "Universe();\n");
+        sb.append("    " + metaModelName.toLowerCase() + "Universe.newDimension(function(dimension){\n");
+        sb.append("        var " + metaModelName.toLowerCase() + "View = dimension.time(0);\n");
         sb.append("\n");
-        sb.append("        //create your root element from '"+metaModelName.toLowerCase()+"View' ::  var element = "+metaModelName.toLowerCase()+"View.create[...]();\n");
-        sb.append("        //set your root element (if necessary) '"+metaModelName.toLowerCase()+"View' ::  "+metaModelName.toLowerCase()+"View.setRoot(<element>);\n");
+        sb.append("        //create your root element from '" + metaModelName.toLowerCase() + "View' ::  var element = " + metaModelName.toLowerCase() + "View.create[...]();\n");
+        sb.append("        //set your root element (if necessary) '" + metaModelName.toLowerCase() + "View' ::  " + metaModelName.toLowerCase() + "View.setRoot(<element>);\n");
         sb.append("\n");
         sb.append("        //Load from JSON\n");
-        sb.append("        "+metaModelName.toLowerCase()+"View.json().load(\"Model_As_String\", function(error){console.error(error);});\n");
+        sb.append("        " + metaModelName.toLowerCase() + "View.json().load(\"Model_As_String\", function(error){console.error(error);});\n");
         sb.append("\n");
         sb.append("        /*  Serialize in JSON:\n");
         sb.append("\n");
-        sb.append("         "+metaModelName.toLowerCase()+"View.json().save(<model>, function(serializedModel, error){\n");
+        sb.append("         //" + metaModelName.toLowerCase() + "View.json().save(<model>, function(serializedModel, error){\n");
         sb.append("         if(!error) {\n");
         sb.append("             //Do something with 'serializedModel'\n");
         sb.append("         } else {\n");
@@ -53,7 +50,9 @@ public class HtmlTemplateGenerator {
         sb.append("</body>\n");
         sb.append("</html>");
 
-        Files.write(Paths.get(targetDir.toAbsolutePath().toString(),"index.html"), sb.toString().getBytes());
+        if (!Paths.get(targetDir.toAbsolutePath().toString(), "index.html").toFile().exists()) {
+            Files.write(Paths.get(targetDir.toAbsolutePath().toString(), "index.html"), sb.toString().getBytes());
+        }
 
     }
